Bugfix + css preparation
fixed a bug where the login session would behave differently based on the localstorage variables.

as of now, state of auth.token is null if there is no local storage equivalent, but it sometimes is a string and sometimes a null object. no login session is present if the auth.token is either null object or 'null' string