Fixed the previous artifact
Now the reloading resides within the userservice.

Also, it's no longer a reload, it now calls the homepage. This makes sense, as it should happen every login / register / logout: it refreshes the top components as well as transitions to the home page, which would have happened even without refreshing the components.

Doing this spares us from streaming the events all the way up to the nav bar from the login, logout and register components.